Update patch set 5
Patch Set 5: Code-Review-1

In the interest of having deployable, testable code in this repository, I still think this commit of the the data stream rest proxy service should add a file to the docs/ area with details about installing, deploying and/or configuring this service so that any community member can stand it up.  Please consider it, thank you.

Patch-set: 5
Label: Code-Review=-1